Update: Animations implemented, with bugs
Currently known bugs:
PlayerDamage does not seem to play, nor does the Damage state seem to be entered whenever the player takes damage.
Whenever the player attacks, if the player attacks with the same attack button during the animation the game will lock on the Attack state and become unplayable.